+++ resources/views/partials/chest.blade.php
@@ -24,7 +24,7 @@
             <button class="btn btn-outline-dark btn-sm dropdown-toggle" type="button" id="dropdownMenuButton" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
                 {{ __('More') }}
             </button>
-            <div class="dropdown-menu" aria-labelledby="dropdownMenuButton">
+            <div class="dropdown-menu dropdown-menu-right" aria-labelledby="dropdownMenuButton">
                 <a class="dropdown-item" href="{{ $chest->permalink }}">{{ __('Permalink') }}</a>
                 @if(auth()->check())
                 <h6 class="dropdown-header">{{ __('Manage') }}</h6>